puzzle logic
added check for if a card is placed in a zone already. as well made it so that when a card is placed in each spot no more cards can be placed.
this is wrong and needs to be changed to just check the solution when theres a card in each space then provided some from of reponse then alow for cahnges to be made.