puzzle logic complete
puzzle now works you can selecte a plant card it then takes you to the puzzle were you can attempt the puzzle. then upon being solved it takes you back to the plant slection with the just solved plant not being an option to chosse.

only part of puzzle left is making it so you can brag wrong cards back to there starting position to change what card is in the slot